refactor: testsのディレクトリ名を変更
