+++ src/features/editor/extensions/nodes/code-block-container/diff-code-block/__tests__/key.browser.spec.ts
@@ -8,9 +8,9 @@ import { waitSelectionChange } from "@/tests/dom";
 import { renderTiptapEditor } from "@/tests/editor";
 import { CodeBlock } from "../../code-block";
 import { CodeBlockFileName } from "../../code-block-file-name";
-import { DiffCodeBlock } from "../../diff-code-block";
-import { DiffCodeLine } from "../../diff-code-block/diff-code-line";
 import { CodeBlockContainer } from "../../index";
+import { DiffCodeBlock } from "..";
+import { DiffCodeLine } from "../diff-code-line";
 
 const baseExtensions = [
   Document,
