+++ src/features/editor/extensions/nodes/code-block-container/diff-code-block/__tests__/markdown.spec.ts
@@ -8,9 +8,9 @@ import { markdownSerializer } from "@/features/editor/lib/to-markdown";
 import { createEditorInstance } from "@/tests/editor-instance";
 import { CodeBlock } from "../../code-block";
 import { CodeBlockFileName } from "../../code-block-file-name";
-import { DiffCodeBlock } from "../../diff-code-block";
-import { DiffCodeLine } from "../../diff-code-block/diff-code-line";
 import { CodeBlockContainer } from "../../index";
+import { DiffCodeBlock } from "..";
+import { DiffCodeLine } from "../diff-code-line";
 
 const baseExtensions = [
   Document,
