+++ src/features/editor/extensions/nodes/code-block-container/diff-code-block/__tests__/parse.spec.ts
@@ -6,9 +6,9 @@ import { describe, expect, it } from "vitest";
 import { createEditorInstance } from "@/tests/editor-instance";
 import { CodeBlock } from "../../code-block";
 import { CodeBlockFileName } from "../../code-block-file-name";
-import { DiffCodeBlock } from "../../diff-code-block";
-import { DiffCodeLine } from "../../diff-code-block/diff-code-line";
 import { CodeBlockContainer } from "../../index";
+import { DiffCodeBlock } from "..";
+import { DiffCodeLine } from "../diff-code-line";
 
 const baseExtensions = [
   Document,
