+++ src/features/editor/extensions/nodes/code-block-container/diff-code-block/__tests__/render.spec.ts
@@ -7,8 +7,8 @@ import { createEditorInstance } from "@/tests/editor-instance";
 import { CodeBlockContainer } from "../..";
 import { CodeBlock } from "../../code-block";
 import { CodeBlockFileName } from "../../code-block-file-name";
-import { DiffCodeBlock } from "../../diff-code-block";
-import { DiffCodeLine } from "../../diff-code-block/diff-code-line";
+import { DiffCodeBlock } from "..";
+import { DiffCodeLine } from "../diff-code-line";
 
 const baseExtensions = [
   Document,
