+++ src/features/editor/extensions/nodes/details/__tests__/command.spec.ts
