+++ src/features/editor/extensions/nodes/details/__tests__/open.browser.spec.ts
