+++ src/features/editor/extensions/nodes/details/__tests__/render.spec.ts
