+++ src/features/editor/extensions/nodes/figure/__tests__/command.spec.ts
