+++ src/features/editor/extensions/nodes/figure/__tests__/input-rule.browser.spec.ts
