+++ src/features/editor/extensions/nodes/figure/__tests__/key.browser.spec.ts
