+++ src/features/editor/extensions/nodes/figure/__tests__/paste.browser.spec.ts
