+++ src/features/editor/extensions/nodes/figure/__tests__/render.spec.ts
